refactor: dedicate place for business types
