+++ src/app/user/wiki/page.tsx
@@ -2,7 +2,7 @@
 
 import { useSession } from 'next-auth/react';
 import { useState, useEffect, useCallback } from 'react';
-import { Allergen, PAP } from '@/lib/schema';
+import { Allergen, PAP } from '@/modules/business-types';
 import { ObjectId } from 'mongodb';
 
 /**
